+++ src/modules/containers.nix
@@ -9,18 +9,21 @@ let
   envContainerName = builtins.getEnv "DEVENV_CONTAINER";
   projectRoot = builtins.path { path = self; name = "source"; };
 
-  nix2containerInput = config.lib.getInput {
-    name = "nix2container";
-    url = "github:nlewo/nix2container";
-    attribute = "containers";
-    follows = [ "nixpkgs" ];
-  };
-  nix2container = nix2containerInput.packages.${pkgs.stdenv.system};
-  mk-shell-bin = config.lib.getInput {
-    name = "mk-shell-bin";
-    url = "github:rrbutani/nix-mk-shell-bin";
-    attribute = "containers";
-  };
+  requiredInputs = config.lib.getInputs [
+    {
+      name = "nix2container";
+      url = "github:nlewo/nix2container";
+      attribute = "containers";
+      follows = [ "nixpkgs" ];
+    }
+    {
+      name = "mk-shell-bin";
+      url = "github:rrbutani/nix-mk-shell-bin";
+      attribute = "containers";
+    }
+  ];
+  nix2container = requiredInputs.nix2container.packages.${pkgs.stdenv.system};
+  mk-shell-bin = requiredInputs.mk-shell-bin;
   shell = mk-shell-bin.lib.mkShellBin { drv = config.shell; nixpkgs = pkgs; };
   bash = "${pkgs.bashInteractive}/bin/bash";
   mkEntrypoint = cfg: pkgs.writeScript "entrypoint" ''